+++ src/randdev.hpp
@@ -4,27 +4,27 @@
 
 #include "prototypes.hpp"
 
+//Qt Includes.
+#include <QtCore>
+
 //C++/11 Standard Headers
 #include <random>
 
 class RandDev
 {
     static RandDev* device;
-    static std::mt19937 randDevice;
+    static std::mt19937_64 randDevice;
 
     public:
         RandDev();
         ~RandDev();
 
         static RandDev* getDevice();
-        static unsigned int getSeed();
+        static qint64 getSeed();
 
         template<typename T>
         T genRandNum(T min, T max)
         {
-            //Re-Seed the device on every use.
-            randDevice.seed( getSeed() );
-
             std::uniform_int_distribution<T> randInt( min, max );
             return randInt( randDevice );
         }