+++ src/upnp.hpp
@@ -52,7 +52,7 @@ class UPNP : public QObject
 
     private:
         void getExternalIP();
-        void extractExternalIP(const QString& message);
+        void extractExternalIP(const QString& action, const QString& message);
         void postSOAP(const QString& action, const QString& message,
                       const QString& protocol, const qint32& port = 0);
         void extractError(const QString& message, const qint32& port,